Add. 2nd implementation

+++ src/main/java/com/thoughtworks/rectangle/Rectangle.java
@@ -1,15 +1,15 @@
 package com.thoughtworks.rectangle;
 class Rectangle{
 
-    private final int length,breadth;
+    private final double length,breadth;
 
-    public Rectangle(int length, int breadth)
+    public Rectangle(double length, double breadth)
     {
       this.length=length;
       this.breadth=breadth;
     }
    
-    public int area(){
+    public double area(){
      return length*breadth;
     }
 }
